Refactor: SingleModeResult 페이지 로직 분리 및 싱글모드 관련 파일들 경로 변경
