+++ src/components/game/singlemode/hooks/useAIPrediction.ts
