+++ src/components/game/singlemode/hooks/useNavigationEffects.ts
