+++ src/components/game/singlemode/hooks/usePageProtection.ts
