+++ src/components/game/singlemode/hooks/useTopicInitialization.ts
@@ -1,5 +1,5 @@
 import { useEffect } from "react";
-import { useDrawingStore } from "../../../stores/drawingStore";
+import { useDrawingStore } from "../../../../stores/drawingStore";
 
 export const useTopicInitialization = () => {
   const { getRandomTopic } = useDrawingStore();
