+++ src/pages/singlemode/AiAnswering.tsx
@@ -2,28 +2,31 @@ import { useState } from "react";
 import { useDrawingStore } from "../../stores/drawingStore";
 import { useGameTimerStore } from "../../stores/gameTimerStore";
 import { useAuthStore } from "../../stores/authStore";
-import { useImageFetch } from "./hooks/useImageFetch";
-import { useAIPrediction } from "./hooks/useAIPrediction";
-import { useNavigationEffects } from "./hooks/useNavigationEffects";
+import { useImageFetch } from "../../components/game/singlemode/hooks/useImageFetch";
+import { useAIPrediction } from "../../components/game/singlemode/hooks/useAIPrediction";
+import { useNavigationEffects } from "../../components/game/singlemode/hooks/useNavigationEffects";
 import SingleModeHeader from "../../components/game/SingleModeHeader";
 import SneakyCat from "../../components/game/SneakyCat";
-import AiResultDisplay from "./AiResultDisplay";
-import AiLoadingDisplay from "./AiLoadingDisplay";
+import AiResultDisplay from "../../components/game/singlemode/AiResultDisplay";
+import AiLoadingDisplay from "../../components/game/singlemode/AiLoadingDisplay";
 
 export default function AiAnswering() {
   const { currentTopic, setAiAnswer, filename } = useDrawingStore();
   const { timeLeft } = useGameTimerStore();
   const user = useAuthStore((state) => state.user);
-  
+
   const [isError, setIsError] = useState(false);
   const [imageReady, setImageReady] = useState(false);
 
   const { imageUrl } = useImageFetch(filename, user?.id, setIsError);
 
-  const { 
-    prediction, 
-    imgRef
-  } = useAIPrediction(imageUrl, currentTopic, setAiAnswer, setIsError, imageReady);
+  const { prediction, imgRef } = useAIPrediction(
+    imageUrl,
+    currentTopic,
+    setAiAnswer,
+    setIsError,
+    imageReady
+  );
 
   useNavigationEffects(prediction, timeLeft, setIsError);
